api/pyproject.toml
[project]
name = "grevia"
version = "0.1.0"
description = "ESG enterprise software"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "sqlalchemy",
    "pymysql",
    "cryptography",
    "anthropic",
    "boto3",
    "python-multipart",
    "pymupdf",
    "python-docx",
    "openpyxl",
    "pillow",
    "pandas",
    "python-dotenv",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "msal",
    "alembic",
    "openai>=2.32.0",
    "langgraph>=0.2.0",
    "httpx>=0.27.0",
    "serpapi>=0.1.5",
    "sse-starlette>=2.1.0",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
    "trustcall>=0.0.39",
    "structlog>=24.1.0",
    "langchain>=1.2.15",
    "langchain-openai>=1.2.1",
    "langsmith>=0.1.0",
]

[project.scripts]
start = "uvicorn main:app --reload"
test = "pytest"
"migrate:up" = "alembic upgrade head"

# Since uv doesn't have 'composite', use a shell-style chain for build
# Or just run the commands sequentially
build = "python setup.py build"

[tool.uv]
managed = true
package = false
